+++ packages/expo-router/src/native-tabs/index.ts
@@ -3,7 +3,6 @@ export type {
   NativeTabsProps,
   NativeTabTriggerProps,
   NativeTabsLabelStyle,
-  NativeTabsActiveStyleType,
   SymbolOrImageSource,
   NativeTabsTabBarItemLabelVisibilityMode,
   NativeTabsBlurEffect,